requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
jax==0.4.8
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==15.0.6.1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
ml-dtypes==0.1.0
mnist==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfunctional==1.4.3
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-version==0.0.2
pytz==2022.7.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn-utils==0.0.15
statsmodels==0.13.5
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.0